[FEAT] member 도메인 테스트 - remove() 함수 테스트 추가
